+++ source/tools/guied/GEApp.cpp
@@ -347,7 +347,7 @@ LRESULT CALLBACK rvGEApp::FrameWndProc ( HWND hWnd, UINT uMsg, WPARAM wParam, LP
 
 			assert ( app );
 			
-			SetWindowLong ( hWnd, GWL_USERDATA, (LONG)app );
+			SetWindowLong ( hWnd, GWL_USERDATA, (intptr_t)app );
 
 			app->mMDIFrame = hWnd;
 			
@@ -364,9 +364,9 @@ LRESULT CALLBACK rvGEApp::FrameWndProc ( HWND hWnd, UINT uMsg, WPARAM wParam, LP
 			app->mToolWindows.Append ( app->mProperties.GetWindow ( ) );
 			app->mToolWindows.Append ( app->mTransformer.GetWindow ( ) );
 			
-			SendMessage ( app->mNavigator.GetWindow ( ), WM_NCACTIVATE, true, (LONG)-1 );
-			SendMessage ( app->mProperties.GetWindow ( ), WM_NCACTIVATE, true, (LONG)-1 );
-			SendMessage ( app->mTransformer.GetWindow ( ), WM_NCACTIVATE, true, (LONG)-1 );			
+			SendMessage ( app->mNavigator.GetWindow ( ), WM_NCACTIVATE, true, (intptr_t)-1 );
+			SendMessage ( app->mProperties.GetWindow ( ), WM_NCACTIVATE, true, (intptr_t)-1 );
+			SendMessage ( app->mTransformer.GetWindow ( ), WM_NCACTIVATE, true, (intptr_t)-1 );
 
 			break;
 		}
@@ -1361,7 +1361,7 @@ int	rvGEApp::ToolWindowActivate ( HWND hwnd, UINT msg, WPARAM wParam, LPARAM lPa
 		{
 			if ( mToolWindows[i] != hwnd &&	mToolWindows[i] != (HWND) lParam )
 			{
-				SendMessage ( mToolWindows[i], WM_NCACTIVATE, keepActive, (LONG)-1 );
+				SendMessage ( mToolWindows[i], WM_NCACTIVATE, keepActive, (intptr_t)-1 );
 			}
 		}
 	}
