+++ source/tools/guied/GEWorkspace.cpp
@@ -143,7 +143,7 @@ bool rvGEWorkspace::Attach ( HWND wnd )
 
 	// Jam the workspace pointer into the userdata window long so 
 	// we can retrieve the workspace from the window later
-	SetWindowLong ( mWnd, GWL_USERDATA, (LONG) this );
+	SetWindowLong ( mWnd, GWL_USERDATA, (intptr_t) this );
 	
 	UpdateTitle ( );
 	
@@ -886,7 +886,7 @@ int	rvGEWorkspace::HandleRButtonDown ( WPARAM wParam, LPARAM lParam )
 		AppendMenu ( popup, MF_STRING|MF_ENABLED|(wrapper->IsSelected()?MF_CHECKED:0), ID_GUIED_SELECT_FIRST + i, mSelectMenu[i]->GetName() );
 	}
 
-	InsertMenu ( menu, 1, MF_POPUP|MF_BYPOSITION, (LONG) popup, "Select" );
+	InsertMenu ( menu, 1, MF_POPUP|MF_BYPOSITION, (intptr_t) popup, "Select" );
 
 	// Bring up the popup menu
 	ClientToScreen ( mWnd, &point );
